+++ src/test/java/com/github/spotbugs/SpotBugsPluginTest.java
@@ -28,14 +28,12 @@ public class SpotBugsPluginTest extends Assert{
   @Rule
   public TemporaryFolder folder = new TemporaryFolder();
 
-  private File sourceDir;
-
   @Before
   public void createProject() throws IOException {
     Files.copy(Paths.get("src/test/resources/SpotBugsPlugin.gradle"), folder.getRoot().toPath().resolve("build.gradle"),
             StandardCopyOption.COPY_ATTRIBUTES);
 
-    sourceDir = folder.newFolder("src", "main", "java");
+    File sourceDir = folder.newFolder("src", "main", "java");
     File to = new File(sourceDir, "Foo.java");
     File from = new File("src/test/java/com/github/spotbugs/Foo.java");
     Files.copy(from.toPath(), to.toPath(), StandardCopyOption.COPY_ATTRIBUTES);
@@ -81,38 +79,6 @@ public class SpotBugsPluginTest extends Assert{
   }
 
   @Test
-  public void testSpotBugsTaskCanFailTheBuild() throws IOException {
-    addClassWithBug();
-
-    BuildResult result = GradleRunner.create()
-        .withProjectDir(folder.getRoot())
-        .withArguments(Arrays.asList("compileJava", "spotbugsMain"))
-        .withPluginClasspath()
-        .buildAndFail(); //Bar.java's bug _should_ fail the build
-
-    Optional<BuildTask> spotbugsMain = findTask(result, ":spotbugsMain");
-    assertTrue(spotbugsMain.isPresent());
-    assertThat(spotbugsMain.get().getOutcome(), is(TaskOutcome.FAILED));
-    assertTrue(result.getOutput().contains("SpotBugs rule violations were found."));
-  }
-
-  @Test
-  public void testSpotBugsTaskWarnsWhenIgnoringFailures() throws IOException {
-    addClassWithBug();
-
-    BuildResult result = GradleRunner.create()
-        .withProjectDir(folder.getRoot())
-        .withArguments(Arrays.asList("compileJava", "spotbugsMain", "-PignoreFailures=true"))
-        .withPluginClasspath()
-        .build(); //Bar.java's bug _should_ fail the build, but logs a warning in this case since we set ignoreFailures = true
-
-    Optional<BuildTask> spotbugsMain = findTask(result, ":spotbugsMain");
-    assertTrue(spotbugsMain.isPresent());
-    assertThat(spotbugsMain.get().getOutcome(), is(TaskOutcome.SUCCESS));
-    assertTrue(result.getOutput().contains("SpotBugs rule violations were found."));
-  }
-
-  @Test
   public void testSpotBugsTestTaskCanRun() throws Exception {
     BuildResult result = GradleRunner.create()
             .withProjectDir(folder.getRoot())
@@ -139,16 +105,6 @@ public class SpotBugsPluginTest extends Assert{
           .findAny();
   }
 
-  private void addClassWithBug() {
-    try {
-      File to = new File(sourceDir, "Bar.java");
-      File from = new File(getClass().getClassLoader().getResource("com/github/spotbugs/Bar.java").getFile());
-      Files.copy(from.toPath(), to.toPath(), StandardCopyOption.COPY_ATTRIBUTES);
-    } catch (IOException e) {
-      throw new RuntimeException("Error adding Bar.java", e);
-    }
-  }
-
   @Test
   public void testLoadToolVersion() {
     SpotBugsPlugin plugin = new SpotBugsPlugin();
@@ -167,21 +123,33 @@ public class SpotBugsPluginTest extends Assert{
         plugin.verifyGradleVersion(GradleVersion.version("3.0"));
     }
 
-    @Test
+    @Test(expected = IllegalArgumentException.class)
     public void testVersionVerifyForGradleVersion4() {
         SpotBugsPlugin plugin = new SpotBugsPlugin();
         plugin.verifyGradleVersion(GradleVersion.version("4.0"));
     }
 
-    @Test
+    @Test(expected = IllegalArgumentException.class)
     public void testVersionVerifyForGradleVersion41() {
         SpotBugsPlugin plugin = new SpotBugsPlugin();
         plugin.verifyGradleVersion(GradleVersion.version("4.1"));
     }
 
-    @Test
+    @Test(expected = IllegalArgumentException.class)
     public void testVersionVerifyForGradleVersion42() {
         SpotBugsPlugin plugin = new SpotBugsPlugin();
         plugin.verifyGradleVersion(GradleVersion.version("4.2"));
+    }
+
+    @Test
+    public void testVersionVerifyForGradleVersion5() {
+        SpotBugsPlugin plugin = new SpotBugsPlugin();
+        plugin.verifyGradleVersion(GradleVersion.version("5.0"));
+    }
+
+    @Test
+    public void testVersionVerifyForGradleVersion51() {
+        SpotBugsPlugin plugin = new SpotBugsPlugin();
+        plugin.verifyGradleVersion(GradleVersion.version("5.1"));
     }
 }